+++ storetests/gs/gsstore_test.go
@@ -1,4 +1,4 @@
-package storetests
+package gs
 
 import (
 	"context"
@@ -12,39 +12,44 @@ import (
 
 	"cloud.google.com/go/storage"
 	"github.com/streamingfast/dstore"
+	"github.com/streamingfast/dstore/storetests"
+	"github.com/streamingfast/logging"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	"google.golang.org/api/iterator"
 )
 
+var zlog, tracer = logging.PackageLogger("dstore", "github.com/streamingfast/dstore/storetests/gs")
+
 // For dfusers, one can use:
+//
 //	STORETESTS_GS_STORE_URL="gs://dfuse-developement-random/store-tests"
-var gsstoreBaseURL = os.Getenv("STORETESTS_GS_STORE_URL")
+var gsStoreBaseURL = os.Getenv("STORETESTS_GS_STORE_URL")
 
 func TestGSStore(t *testing.T) {
-	if gsstoreBaseURL == "" {
+	if gsStoreBaseURL == "" {
 		t.Skip("You must provide a valid Google Storage Bucket via STORETESTS_GS_STORE_URL environment variable to execute those tests")
 		return
 	}
 
-	TestAll(t, createGSStoreFactory(t, gsstoreBaseURL, "", false))
+	storetests.TestAll(t, createGSStoreFactory(t, gsStoreBaseURL, "", false))
 }
 
 func TestGSStore_Overwrite(t *testing.T) {
-	if gsstoreBaseURL == "" {
+	if gsStoreBaseURL == "" {
 		t.Skip("You must provide a valid Google Storage Bucket via STORETESTS_GS_STORE_URL environment variable to execute those tests")
 		return
 	}
 
-	TestAll(t, createGSStoreFactory(t, gsstoreBaseURL, "", true))
+	storetests.TestAll(t, createGSStoreFactory(t, gsStoreBaseURL, "", true))
 }
 
-func createGSStoreFactory(t *testing.T, directory string, compression string, overwrite bool) StoreFactory {
+func createGSStoreFactory(t *testing.T, directory string, compression string, overwrite bool) storetests.StoreFactory {
 	random := rand.NewSource(time.Now().UnixNano())
 
-	return func() (dstore.Store, StoreCleanup) {
+	return func() (dstore.Store, storetests.StoreCleanup) {
 		testPath := fmt.Sprintf("dstore-gsstore-tests-%08x", random.Int63())
-		fullPath := gsstoreBaseURL
+		fullPath := gsStoreBaseURL
 		if !strings.HasSuffix(fullPath, "/") {
 			fullPath += "/"
 		}
@@ -60,7 +65,7 @@ func createGSStoreFactory(t *testing.T, directory string, compression string, ov
 		require.NoError(t, err)
 
 		return store, func() {
-			if noCleanup {
+			if storetests.NoCleanup {
 				client.Close()
 				return
 			}
@@ -70,7 +75,7 @@ func createGSStoreFactory(t *testing.T, directory string, compression string, ov
 				Prefix: strings.TrimLeft(storeURL.Path, "/") + "/",
 			})
 
-			if traceEnabled {
+			if tracer.Enabled() {
 				zlog.Debug("cleaning out bucket", zap.String("bucket", storeURL.Host), zap.String("prefix", storeURL.Path))
 			}
 
@@ -83,7 +88,7 @@ func createGSStoreFactory(t *testing.T, directory string, compression string, ov
 				require.NoError(t, err)
 				object := bucket.Object(value.Name)
 
-				if traceEnabled {
+				if tracer.Enabled() {
 					zlog.Debug("about to delete bucket file", zap.String("name", value.Name))
 				}
 				err = object.Delete(context.Background())