+++ storetests/logging.go
@@ -1,7 +1,0 @@
-package storetests
-
-import (
-	"github.com/streamingfast/logging"
-)
-
-var zlog, _ = logging.PackageLogger("dstore", "github.com/streamingfast/dstore/storetests")