+++ storetests/mock/init_test.go
@@ -1,0 +1,7 @@
+package mock
+
+import "github.com/streamingfast/logging"
+
+func init() {
+	logging.InstantiateLoggers()
+}