+++ storetests/s3/init_test.go
@@ -1,0 +1,7 @@
+package s3
+
+import "github.com/streamingfast/logging"
+
+func init() {
+	logging.InstantiateLoggers()
+}